Need to finish all undos, and main

+++ cs4773-hw3/src/main/java/commands/shapes/Rectangle.java
@@ -1,29 +1,29 @@
 package commands.shapes;
 
 public class Rectangle extends Shape {
-    private int height;
-    private int width;
+    private double height;
+    private double width;
     private double[] cooridnates = origin.getOrigin();
 
-    public Rectangle(int height, int width) {
+    public Rectangle(double height, double width) {
         this.height = height;
         this.width = width;
         color.setColor("Red");
     }
 
-    public void setHeight(int height) {
+    public void setHeight(double height) {
         this.height = height;
     }
 
-    public void setWidth(int width) {
+    public void setWidth(double width) {
         this.width = width;
     }
 
-    public int getHeigth() {
+    public double getHeigth() {
         return height;
     }
 
-    public int getWidth() {
+    public double getWidth() {
         return width;
     }
